+++ Example/AELogDemo iOS/ViewController.swift
@@ -11,31 +11,41 @@ class ViewController: UIViewController {
     
     override func viewDidLoad() {
         super.viewDidLoad()
-        log()
+        logToDebugger()
     }
     
     override func viewWillAppear(_ animated: Bool) {
         super.viewWillAppear(animated)
-        log()
+        logToDebugger()
     }
     
     override func viewDidAppear(_ animated: Bool) {
         super.viewDidAppear(animated)
-        log()
+        logToDebugger()
 
+        let text = "One two three four five"
         let x: CGFloat = 21
         let y: CGFloat = 8
         let size = CGSize(width: 19, height: 84)
         let rect = CGRect(x: x, y: y, width: size.width, height: size.height)
-        log(elements: x, y, size, rect)
+        let range = 1...5
+        logToDebugger(items: text, x, y, size, rect, range, Log.shared, self)
+
+        queue.async {
+            logToDebugger("This is coming from background thread.")
+        }
+
+        logToDevice("This will be logged to device console.")
     }
+
+    private let queue = DispatchQueue(label: "Custom")
     
     @IBAction func didTapButton(_ sender: UIButton) {
         let queue = DispatchQueue.global()
         queue.async {
             generateLogLines(count: Int.random(max: 1000))
             DispatchQueue.main.async(execute: {
-                log(message: sender)
+                logToDebugger(items: sender)
             })
         }
     }